requirements.txt
python-dotenv
requests
pandas
numpy
python-dotenv
ccxt
matplotlib
